requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.6.0
torchvision==0.21.0
pydantic==2.5.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart==0.0.6 